moved OntologyRelationship from read package to one level higher, where the graph model resides
