+++ ontology-rest/src/main/java/uk/ac/ebi/quickgo/ontology/traversal/read/OntologyGraphConfig.java
@@ -1,6 +1,7 @@
 package uk.ac.ebi.quickgo.ontology.traversal.read;
 
 import uk.ac.ebi.quickgo.ontology.traversal.OntologyGraph;
+import uk.ac.ebi.quickgo.ontology.traversal.OntologyRelationship;
 
 import java.io.IOException;
 import java.util.ArrayList;
