+++ ontology-rest/src/test/java/uk/ac/ebi/quickgo/ontology/traversal/read/OntologyRelationshipValidatorTest.java
@@ -1,6 +1,7 @@
 package uk.ac.ebi.quickgo.ontology.traversal.read;
 
 import uk.ac.ebi.quickgo.ontology.traversal.OntologyRelationType;
+import uk.ac.ebi.quickgo.ontology.traversal.OntologyRelationship;
 
 import org.junit.Before;
 import org.junit.Test;
